changed imports to point at imported oz lib

+++ src/IFUM.sol
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 pragma solidity ^0.8.0;
 
-import "acc-erc20/contracts/IERC20.sol";
+import "lib/openzeppelin-contracts/contracts/token/ERC20/IERC20.sol";
 
 interface IFUM is IERC20 {
     /**
@@ -10,12 +10,12 @@ interface IFUM is IERC20 {
      * @param recipient address to mint to
      * @param amount amount to mint
      */
-    function mint(address recipient, uint amount) external;
+    function mint(address recipient, uint256 amount) external;
     /**
      * @notice Burn FUM from holder
      *
      * @param holder address to burn from
      * @param amount amount to burn
      */
-    function burn(address holder, uint amount) external;
+    function burn(address holder, uint256 amount) external;
 }
